requirements.txt
fastapi
uvicorn
serpapi
python-jobspy
pandas
python-dotenv